+++ src/index.ts
@@ -1,3 +1,5 @@
 export { IndicatorScrollView } from './IndicatorScrollView';
 export type { IndicatorScrollViewProps } from './IndicatorScrollView';
+export { PaginatedIndicatorScrollView } from './PaginatedIndicatorScrollView';
+export type { PaginatedIndicatorScrollViewProps, MenuItem } from './PaginatedIndicatorScrollView';